pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "catflow-service-detector"
dynamic = ["version"]
readme = "README.md"
description = "Detection/notification service for an object recognition pipeline"
authors = [{ name = "Ian Kilgore", email = "iank@iank.org"}]
license = {file = "LICENSE"}
requires-python = ">=3.7"
dependencies = [
  "catflow-worker>=0.1.7",
  "opencv-python-headless",
  "aiohttp",
  "numpy",
  "scipy",
  "pillow",
]

[tool.setuptools]
packages = ["catflow_service_detector"]

[project.scripts]
catflow-service-detector = "catflow_service_detector:main"

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "moto",
  "boto3",
]

[tool.setuptools_scm]
write_to = "catflow_service_detector/_version.py"

[tool.pytest.ini_options]
filterwarnings = [
    # Fail on any warnings that are not explicitly matched below
    "error",
]
